requirements.txt
# Core dependencies
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
numpy>=1.24.0
Werkzeug>=2.3.0
gunicorn>=21.0.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.2.0

# Code quality
flake8>=6.1.0
black>=23.0.0
isort>=5.12.0
